Refactor != operator overload.
Remove some unnecessary condition evaluations. The old version checked
to make sure that curr was between first and last, but this was
unnecessary since curr must be "after" first. All that needs to be
checked is that last hasn't been passed, which is all that is checked
now.

Further, we're only going to use an if-else instead of a returning the
result of a single expression, this will allow us to perform a
comparison with last only once.